refactor: Update City and Coordonates classes to improve data types and add us_capitals data file

+++ TP4/src/model/Coordonates.java
@@ -9,25 +9,25 @@ package model;
 
 public class Coordonates {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Coordonates(int x, int y) {
+    public Coordonates(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int x() {
+    public double x() {
         return x;
     }
 
-    public int y() {
+    public double y() {
         return y;
     }
 
     public double distanceTo(Coordonates other) {
-        int dx = this.x - other.x;
-        int dy = this.y - other.y;
+        double dx = this.x - other.x;
+        double dy = this.y - other.y;
         return Math.sqrt(dx * dx + dy * dy);
     }
